% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{get.lam.max2}
\alias{get.lam.max2}
\title{Obtain lambda_max for for generating sequence of tuning parameter in G-CURE problem}
\usage{
get.lam.max2(Y, X, familygroup, offset)
}
\arguments{
\item{Y}{Multivariate response matrix}

\item{X}{Design matrix}

\item{familygroup}{indicator {1,2,3} for {gaussian, binary, poisson}}

\item{offset}{offset term matrix}
}
\description{
Obtain lambda_max for for generating sequence of tuning parameter in G-CURE problem
}
