% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{getKappaC0}
\alias{getKappaC0}
\title{Obtain scaling constant for monotone decreasing G-CURE problem}
\usage{
getKappaC0(X0, familygroup)
}
\arguments{
\item{X0}{Design matrix}

\item{familygroup}{indicator {1,2,3} for {gaussian, binary, poisson}}
}
\description{
Obtain scaling constant for monotone decreasing G-CURE problem
}
