% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{getNullDev}
\alias{getNullDev}
\title{Obtain null deviance to utilize in convergence}
\usage{
getNullDev(Y, ofset, familygroup, naind)
}
\arguments{
\item{Y}{Multivariate response matrix}

\item{ofset}{offset matrix}

\item{familygroup}{indicator {1,2,3} for {gaussian, binary, poisson}}

\item{naind}{index of missing entries}
}
\description{
Obtain null deviance to utilize in convergence
}
