% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{glmCol}
\alias{glmCol}
\title{Fit glm Columnwise on the control variable}
\usage{
glmCol(Y, X0, ofset, family, familygroup, q, cIndex)
}
\arguments{
\item{Y}{Multivariate response matrix}

\item{X0}{design matrix}

\item{ofset}{offset  matrix}

\item{family}{{gaussian, Bernouli, poisson}}

\item{familygroup}{indicator {1,2,3} for {gaussian, binary, poisson}}

\item{q}{number of outcomes}

\item{cIndex}{index of control variable in X0}
}
\description{
Fit glm Columnwise on the control variable
}
