% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{logLikehood}
\alias{logLikehood}
\title{loglikelihood of the observation}
\usage{
logLikehood(Y, MU, Sigma = 1, family)
}
\arguments{
\item{Y}{outcome variables}

\item{MU}{natural parameter matrix}

\item{Sigma}{dispersion parameter for gacussian}

\item{family}{gaussian binomial poisson}
}
\description{
loglikelihood of the observation
}
