% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{objFun5}
\alias{objFun5}
\title{Evaluate of objective function}
\usage{
objFun5(Y, mu, Phi, familygroup)
}
\arguments{
\item{Y}{outcome variables}

\item{mu}{natural parameter matrix}

\item{Phi}{dispersion parameter}

\item{familygroup}{gaussian binomial poisson}
}
\description{
Evaluate of objective function
}
