% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{updateFitObject}
\alias{updateFitObject}
\title{Rescale gaussian response}
\usage{
updateFitObject(fit, mx)
}
\arguments{
\item{fit}{fitted object from gsecure and geecure}

\item{mx}{scaling value}
}
\description{
Rescale gaussian response
}
