% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfun.R
\name{getScaleGaussian}
\alias{getScaleGaussian}
\title{Suitably scale gaussian response for unit variance}
\usage{
getScaleGaussian(Yt, X0, familygroup)
}
\arguments{
\item{X0}{design matrix}

\item{familygroup}{indicator {1,2,3} for {gaussian, binary, poisson}}

\item{Y0}{Multivariate response matrix}
}
\description{
Suitably scale gaussian response for unit variance
}
